Refactor Migration
