+++ BE/Initializers/Database.go
@@ -6,8 +6,6 @@ import (
 	"time"
 	"wan-api-kol-event/Models"
 
-	"math/rand"
-
 	"gorm.io/driver/postgres"
 	"gorm.io/gorm"
 	"gorm.io/gorm/logger"
@@ -40,15 +38,6 @@ func ConnectToDB() {
 	}
 }
 
-func randomString(length int) string {
-	const charset = "abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789"
-	b := make([]byte, length)
-	for i := range b {
-		b[i] = charset[rand.Intn(len(charset))]
-	}
-	return string(b)
-}
-
 func tableExists(db *gorm.DB, tableName string) bool {
 	var count int64
 	db.Table("pg_tables").Where("tablename = ?", tableName).Count(&count)
@@ -62,34 +51,7 @@ func MigrateAndSeed() {
 			log.Fatal("Migration failed:", err)
 		}
 		for i := 0; i < 50; i++ {
-			kol := Models.Kol{
-				UserProfileID:        int64(i + 1),
-				Language:             randomString(2),
-				Education:            randomString(10),
-				ExpectedSalary:       rand.Int63n(10000) + 3000,
-				ExpectedSalaryEnable: rand.Intn(2) == 1,
-				ChannelSettingTypeID: int64(rand.Intn(5) + 1),
-				IDFrontURL:           "http://example.com/idfront/" + randomString(3) + ".jpg",
-				IDBackURL:            "http://example.com/idback/" + randomString(3) + ".jpg",
-				PortraitURL:          "http://example.com/portrait/" + randomString(3) + ".jpg",
-				RewardID:             int64(rand.Intn(5) + 1),
-				PaymentMethodID:      int64(rand.Intn(5) + 1),
-				TestimonialsID:       int64(rand.Intn(5) + 1),
-				VerificationStatus:   rand.Intn(2) == 1,
-				Enabled:              rand.Intn(2) == 1,
-				ActiveDate:           time.Now(),
-				Active:               true,
-				CreatedBy:            "admin",
-				CreatedDate:          time.Now(),
-				ModifiedBy:           "admin",
-				ModifiedDate:         time.Now(),
-				IsRemove:             false,
-				IsOnBoarding:         rand.Intn(2) == 1,
-				Code:                 "KOL" + randomString(3),
-				PortraitRightURL:     "http://example.com/portraitright/" + randomString(3) + ".jpg",
-				PortraitLeftURL:      "http://example.com/portraitleft/" + randomString(3) + ".jpg",
-				LivenessStatus:       rand.Intn(2) == 1,
-			}
+			kol := Models.GetRandomKol(int64(i))
 
 			if err := DB.Create(&kol).Error; err != nil {
 				log.Println("Error inserting fake data:", err)
